ServiceabilityAgent/requirements.txt
# Python dependencies for Serviceability Agent

# Core dependencies
google-adk>=1.20.0
google-genai>=1.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client for GIS API
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
